requirements.txt
asgiref==3.2.3
Brotli==1.0.7
Django==3.0.4
django-appconf==1.0.3
django-ckeditor==5.9.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mdeditor==0.1.17
django-mptt==0.11.0
django-taggit==1.2.0
Markdown==3.2.1
pilkit==2.0
Pillow==7.0.0
Pygments==2.6.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
